version 1.4 algo no anda bien aca

+++ src/main/java/ar/com/Snipplet/persistencia/Persistencia.java
@@ -240,6 +240,7 @@ public class Persistencia {
 
 	public void eliminarYCrearArchivo(String filename) throws IOException {
 		File file = new File(prefix + filename);
+		if(file.exists())
 		file.delete();
 		file.createNewFile();
 	}
